requirements.txt
anyio==4.9.0
APScheduler==3.11.0
asyncpg==0.30.0
certifi==2025.4.26
charset-normalizer==3.4.4
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
isort==6.0.1
nest-asyncio==1.6.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-telegram-bot==22.1
python-telegram-bot-calendar==1.0.5
requests==2.32.5
sentry-sdk==2.29.1
six==1.17.0
sniffio==1.3.1
typing_extensions==4.14.0
tzdata==2025.2
tzlocal==5.3.1
urllib3==2.4.0
